pyproject.toml
[tool.poetry]
name = "lsu-pilot"
version = "0.1.0"
description = ""
authors = ["Ahmad Syafiq <ahmadsyafiqruhazat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.1.1"
openai = "^1.25.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
tiktoken = "^0.6.0"
langchain = "^0.1.17"
numpy = "^1.26.4"
scipy = "^1.13.0"
cairosvg = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
